crypto/requirements.txt
django
djangorestframework
requests
bs4
